move inline code from poly_extra.h to proper implementation files
or delete unused code
