+++ poly_extra/fmpz_poly_evaluate_at_one.c
@@ -1,0 +1,17 @@
+/*
+   Copyright (C) 2016 Vincent Delecroix
+
+    This file is part of e-antic
+
+    e-antic is free software: you can redistribute it and/or modify it under
+    the terms of the GNU Lesser General Public License (LGPL) as published
+    by the Free Software Foundation; either version 3.0 of the License, or
+    (at your option) any later version.  See <http://www.gnu.org/licenses/>.
+*/
+
+#include "../e-antic/poly_extra.h"
+
+void fmpz_poly_evaluate_at_one(fmpz_t res, fmpz * p, slong len)
+{
+    _fmpz_vec_sum(res, p, len);
+}
