+++ poly_extra/fmpz_poly_has_real_root.c
@@ -1,0 +1,18 @@
+/*
+   Copyright (C) 2016 Vincent Delecroix
+
+    This file is part of e-antic
+
+    e-antic is free software: you can redistribute it and/or modify it under
+    the terms of the GNU Lesser General Public License (LGPL) as published
+    by the Free Software Foundation; either version 3.0 of the License, or
+    (at your option) any later version.  See <http://www.gnu.org/licenses/>.
+*/
+
+#include "../e-antic/poly_extra.h"
+
+int fmpz_poly_has_real_root(fmpz_poly_t pol)
+{
+    return _fmpz_poly_has_real_root(pol->coeffs, pol->length);
+}
+
